Updated migrations to fix product price display. Also set up product printprice and cart totalprice class methods, set up display in json. App now displays prices correctly, and adds up a total correctly in the cart.

+++ backend/db/migrate/20200408005043_create_products.rb
@@ -4,7 +4,7 @@ class CreateProducts < ActiveRecord::Migration[6.0]
       t.string :name
       t.string :category
       t.string :description
-      t.decimal :price, precision:2
+      t.integer :price
       t.string :image_url
 
       t.timestamps
